feat: implement background type configuration
You can now configure the card to use an image or person entity profile picture as an
icon alternative.

To do this there is a new Select Box in the Editor with 4 Options:
- No Background
- Color Circle (default)
- Custom Image
- Person Profile Picture

Custom Image allows you to input a path to an image either locally stored in your
/www/ folder or externally.

Profile Picture gives you a Person Selector which then shows the persons profile
picture as card image.

No Background and Color Circle are existing options which just have been renamed.